% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_customer_behavior_table.R
\name{create_customer_behavior_table}
\alias{create_customer_behavior_table}
\title{Title TODO THIS IS OLD FUNCTION FOR DEMO}
\usage{
create_customer_behavior_table(df, org, lookup_vector)
}
\arguments{
\item{df}{Raw data frame from org at transactional level}

\item{org}{Three letter code used to check correct lookup table used}

\item{lookup_vector}{Named vector from global environment in app that provides old and used names}
}
\value{
Customer Behavior Table
}
\description{
Title TODO THIS IS OLD FUNCTION FOR DEMO
}
