% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofs_functions.R
\name{ofs_clean_transaction_order_type}
\alias{ofs_clean_transaction_order_type}
\title{OFS Clean Transactions Table}
\usage{
ofs_clean_transaction_order_type(ofs_transaction_table)
}
\arguments{
\item{ofs_transaction_table}{Cleaned transactions table for OFS}
}
\value{
Table with internal variables standardized
}
\description{
OFS Clean Transactions Table
}
