% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_calculate_mrr.R
\name{toy_calculate_mrr}
\alias{toy_calculate_mrr}
\title{toy_calculate_mrr}
\usage{
toy_calculate_mrr(n_customers, product_price)
}
\arguments{
\item{n_customers}{The Number of Customers}

\item{product_price}{How Much You Might Charge}
}
\value{
ans
}
\description{
toy_calculate_mrr
}
\examples{
toy_calculate_mrr(n_customers = 10, product_price = 300)
}
